6. DE Job vacancies full ETL pipeline (Rapid APi) 1/Job_Scraping/requirements.txt
requests==2.28.1
pandas
psycopg2-binary==2.9.5
sqlalchemy
